Fixed pow2 slots-to-coeffs

+++ src/lib.rs
@@ -37,7 +37,7 @@ macro_rules! ring_literal {
         {
             let ring = $ring;
             let base_ring = ring.base_ring();
-            <_ as feanor_math::rings::extension::FreeAlgebraStore>::from_canonical_basis(&ring, $iter.into_iter().map(|x| <_ as feanor_math::homomorphism::Homomorphism<_, _>>::map(&base_ring.int_hom(), x)))
+            <_ as feanor_math::rings::extension::FreeAlgebraStore>::from_canonical_basis(&ring, $iter.into_iter().map(|x| <_ as feanor_math::homomorphism::Homomorphism<_, _>>::map(&base_ring.int_hom(), x as i32)))
         }
     };
 }
